Join adjacent objects when freeing an object only if its resulted free size is less than ICELLS.
If a free size is greater than ICELLS, it is recognized as a large object.
In this case, malloc()/free() pair is used to deal it. Unfortunately,
however, this is not true. A free size can be grater than ICELLS while
joining adjacents free objects. As a result, this causes a
memory-releated-problem such as a memory corruption and SIGSEGV.