Read in text mode in a needed place only
Converting CR+LF to LF in blocking_read() which is underlying read
function, causes a trouble in c_read().

For example, libtool recognizes a .la file with CR+LF as an invalid
file.

This is because c_read() calls shf_flush() to empty a buffer.
shf_flush() restores a file position according to remaing buffer size.
By the way, the buffer contains converted LF from CR+LF. After all,
a file position is restored at an invalid position, and multiple
calls to c_read() eat some bytes at the begining of a line.

    modified:   c_sh.c
    modified:   io.c
    modified:   misc.c
    modified:   os2/os2.c
    modified:   shf.c