os2: do not filter stdin
Filtering stdin causes some problems.

1. Some programs such as printf.exe never read from stdin. In this case,
writing to a pipe would block after a limit size, 8192 on EMX.

2. dup() shares a file pointer. If reading from fd in a filter
thread, it will be consumed forever. Any other access to that fd cannot
get any contents. In practice, git rebase -i fails due to this. The
sample codes are like this:

    while read ...
    do
        ...
    done < file1 > file2

c_read() implementing read command of pdksh reads 1 line from a handle
which is dup()ed before executing commands in while block. When commands
in while block are executed, a filter thread of stdin will reads from a
dup()ed handle. Because a file pointer is shared between a original
handle and a dup()ed handle, a file pointer will be increased by the
size read in a filter thread. Next read command will read something
wrong.

This approach disables to convert CR+LF to LF from stdin. However, only
one case is not filtered. It is a redirection from a file to stdin. The
above sample codes are good examples.

Any other cases remains not changed. Because when writing to pipes,
CR+LF is converted to LF, it is not necessary to convert CR+LF to LF
when reading.

    modified:   os2/os2.c